+++ src/cli.py
@@ -2,15 +2,11 @@ import argparse
 import os
 import sys
 
-from directory_processors import FlatDirectoryProcessor
-from watermarking import Corner
-
-DARK_WATERMARK_FILEPATH = './watermark-dark.png'
-LIGHT_WATERMARK_FILEPATH = './watermark-light.png'
-DEFAULT_CORNERS = [Corner.UPPER_LEFT, Corner.UPPER_RIGHT, Corner.LOWER_LEFT, Corner.LOWER_RIGHT]
+from core.directory_processors import FlatDirectoryProcessor
 
 
 def main():
+    # TODO: picking corners (now falling back to default)
     parser = argparse.ArgumentParser(description='Adds watermark to each photo in a folder, supports .jpg files')
     parser.add_argument('-f', '--folder', dest='folder', default='.', help='Path to a folder with photos', type=str)
     parser.add_argument('--width', dest='width', default=0.15, help='[0.0, 1.0] Max watermark to image width ratio',
@@ -29,12 +25,9 @@ def main():
         sys.exit()
 
     directory_processor = FlatDirectoryProcessor(
-        dark_watermark_filepath=DARK_WATERMARK_FILEPATH,
-        light_watermark_filepath=LIGHT_WATERMARK_FILEPATH,
         max_width_proportion=args.width,
         max_height_proportion=args.height,
         opacity=args.opacity,
-        corners=DEFAULT_CORNERS
     )
 
     directory_processor.handle_directory(args.folder)
