+++ src/core/watermarking.py
