+++ src/main.py
@@ -1,0 +1,10 @@
+from gui.gui import App
+
+
+def main():
+    app = App()
+    app.mainloop()
+
+
+if __name__ == '__main__':
+    main()
