+++ src/resources/watermarks/__init__.py
@@ -1,0 +1,5 @@
+import os
+
+WATERMARKS_DIR = os.path.abspath('resources/watermarks')
+DEFAULT_LIGHT_WATERMARK = os.path.join(WATERMARKS_DIR, 'watermark-light.png')
+DEFAULT_DARK_WATERMARK = os.path.join(WATERMARKS_DIR, 'watermark-dark.png')
