+++ src/resources/watermarks/watermark-dark.png
