+++ src/resources/watermarks/watermark-light.png
